feat: refactor schedule components
