+++ src/components/schedule/components/Break.tsx
@@ -1,4 +1,4 @@
-import styles from "./index.module.css";
+import styles from "../styles/index.module.css";
 
 export const Break = ({title}: {title: string}) => {
   return (
